# swagger.yml
swagger: "2.0"
info:
  title: Hasker API
  description: |
    Hacker API implementation description
  version: "0.1.0"
schemes:
- http
basePath: /api/v1
consumes:
- application/json
produces:
- application/json
paths:
...
  /questions:
    get:
      summary: Список вопросов
      description: |
        Список вопросов упорядоченых по дате создания.
        По 20 вопросов на страницу.
      operationId: questions-list
      parameters:
      - in: query
        name: page
        type: integer
        minimum: 1
        description: Номер запрашиваемой страницы
      - in: query
        name: sort
        type: string
        enum: [new, popular]
        description: |
          Тип сортировки:
          * new - Список вопросов упорядоченых по дате создания
          * popular - Список вопросов упорядоченых рейтингу
      responses:
        200:
          description: Упорядоченный список вопросов
          schema:
            $ref: '#/definitions/QuestionsList'
        401:
          description: Необходима авторизация
        404:
          description: Не верно указан номер страницы для пагинации.
...
definitions:
...
  QuestionInfo:
    description: Краткая информация о вопросе
    type: object
    properties:
      pk:
        type: number
        format: int64
        description: Идентификатор вопроса.
        readOnly: true
        x-isnullable: false
      title:
        type: string
        format: text
        description: Заголовок вопроса.
        example: To be, or not to be?
        x-isnullable: false
      url:
        type: string
        format: uri
        description: Ссылка на детальную информацию о вопросе.
        x-isnullable: false
  QuestionsList:
    description: Списов вопросов
    type: array
    items:
      $ref: '#/definitions/QuestionInfo'
...